+++ viewer/src/main/kotlin/li/doerf/feeder/viewer/config/JwtTokenFilterConfigurer.kt
@@ -1,4 +1,4 @@
-package li.doerf.feeder.viewer.config.security
+package li.doerf.feeder.viewer.config
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.security.config.annotation.SecurityConfigurerAdapter
