+++ viewer/src/main/kotlin/li/doerf/feeder/viewer/config/JwtTokenProvider.kt
@@ -1,10 +1,11 @@
-package li.doerf.feeder.viewer.config.security
+package li.doerf.feeder.viewer.config
 
 import io.jsonwebtoken.JwtException
 import io.jsonwebtoken.Jwts
 import io.jsonwebtoken.SignatureAlgorithm
 import li.doerf.feeder.viewer.entities.Role
 import li.doerf.feeder.viewer.exception.HttpException
+import li.doerf.feeder.viewer.services.UserDetailsService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.HttpStatus
