+++ viewer/src/main/kotlin/li/doerf/feeder/viewer/config/WebSecurityConfig.kt
@@ -1,8 +1,6 @@
 package li.doerf.feeder.viewer.config
 
 import li.doerf.feeder.common.util.getLogger
-import li.doerf.feeder.viewer.config.security.JwtTokenFilterConfigurer
-import li.doerf.feeder.viewer.config.security.JwtTokenProvider
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
