+++ viewer/src/main/kotlin/li/doerf/feeder/viewer/config/WebSocketConfig.kt
@@ -1,7 +1,6 @@
 package li.doerf.feeder.viewer.config
 
 import li.doerf.feeder.common.util.getLogger
-import li.doerf.feeder.viewer.config.security.JwtTokenProvider
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Configuration
 import org.springframework.messaging.Message
