+++ viewer/src/main/kotlin/li/doerf/feeder/viewer/services/UserService.kt
@@ -1,7 +1,7 @@
 package li.doerf.feeder.viewer.services
 
 import li.doerf.feeder.common.util.getLogger
-import li.doerf.feeder.viewer.config.security.JwtTokenProvider
+import li.doerf.feeder.viewer.config.JwtTokenProvider
 import li.doerf.feeder.viewer.dto.UserResponseDto
 import li.doerf.feeder.viewer.entities.Role
 import li.doerf.feeder.viewer.entities.User
